pyproject.toml
[project]
name = "heartdd"
version = "0.1.0"
description = ""
authors = [{name = "Adam Wozny"},{name = "Michal Czajor"},{name = "Jakub Chrzanowski"}]
license = {text = "MIT"}
dependencies = [
    "grpcio==1.66.2",
    "matplotlib==3.8.4",
    "numpy==2.1.3",
    "opencv-python==4.10.0.84",
    "pandas==2.2.3",
    "Pillow==11.0.0",
    "protobuf==3.19.6",
    "pydicom==3.0.1",
    "scipy==1.14.1",
    "torch==2.5.0",
    "torchvision==0.20.0",
    "tqdm==4.66.5",
    "gdown~=5.2.0",
    "scikit-learn~=1.5.2",
    "scikit-image~=0.24.0"
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
